+++ EssentialApp/EssentialAppTests/Helpers/XCTestCase+FeedImageDataLoader.swift
@@ -1,0 +1,36 @@
+//
+//  XCTestCase+FeedImageDataLoader.swift
+//  EssentialAppTests
+//
+//  Created by Hitesh on 22/02/22.
+//
+
+import XCTest
+import EssentialFeed
+
+protocol FeedImageDataLoaderTestCase: XCTestCase {}
+
+extension FeedImageDataLoaderTestCase {
+     func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
+         let exp = expectation(description: "Wait for load completion")
+
+         _ = sut.loadImageData(from: anyURL()) { receivedResult in
+             switch (receivedResult, expectedResult) {
+             case let (.success(receivedFeed), .success(expectedFeed)):
+                 XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
+
+             case (.failure, .failure):
+                 break
+
+             default:
+                 XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
+             }
+
+             exp.fulfill()
+         }
+
+         action()
+
+         wait(for: [exp], timeout: 1.0)
+     }
+ }
